backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Werkzeug==3.0.1
mysql-connector-python==8.0.26